fix(Queries): double scroll [YTFRONT-5943]
