+++ packages/ui/src/ui/pages/query-tracker/QueriesList/QueriesHistoryList/QueriesHistoryList.scss
@@ -6,8 +6,4 @@
         padding: 12px 12px 0;
         margin-bottom: 8px;
     }
-
-    &__pagination {
-        padding: 8px 0;
-    }
 }
